refactor: create sub namespace for processors
